python/requirements.txt
pySerial
ino
